+++ codegen-activity.yml
@@ -10,4 +10,3 @@ generates:
       - "typescript-resolvers"
 config:
   mappers:
-    Person: "../src/Person/types#PersonType"
