+++ services/activity/src/Event/Event.schema.graphql
