+++ services/activity/src/Group/Group.schema.graphql
