+++ services/activity/src/GroupAward/GroupAward.schema.graphql
