--- services/activity/src/Person/types.ts
@@ -1,3 +1,0 @@
-import { Person } from "../../generated/activity-db";
-
-export type PersonType = Person;
