+++ services/activity/src/School/School.schema.graphql
@@ -1,4 +1,3 @@
 type School @key(fields: "id") {
   id: ID!
-  People: [Person]
 }
