+++ services/activity/src/Venture/Venture.schema.graphql
