+++ services/school/src/Person/types.ts
@@ -1,0 +1,3 @@
+import { Person } from "../../generated/school-db";
+
+export type PersonType = Person;
