+++ services/school/src/Relationship/Relationship.schema.graphql
