+++ services/school/src/School/School.schema.graphql
@@ -5,5 +5,6 @@ type School @key(fields: "id") {
   Addresses: [Address]
   Emails: [Email]
   LegalForms: [LegalForm]
+  People: [Person]
   Phones: [Phone]
 }
